move svg to component
